refactor: Chart 함수 통일 후 utils로 데이터 포맷팅
